Finished redoing dependencies to definitions
Redone ClassDefinition and InvalidDefinition
Removed unused and unfininished NamedDependency and NamedClassDependency
